requirements.txt
DingtalkChatbot
Jinja2
PyYAML
sanic
